Move kv journal stuff into its own module
