+++ bd-resilient-kv/src/kv_journal/double_buffered.rs
@@ -5,7 +5,7 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use crate::kv_journal::KVJournal;
+use super::KVJournal;
 use bd_bonjson::Value;
 use std::collections::HashMap;
