+++ bd-resilient-kv/src/kv_store.rs
@@ -5,8 +5,12 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use crate::kv_journal::{HighWaterMarkCallback, KVJournal};
-use crate::{DoubleBufferedKVJournal, MemMappedKVJournal};
+use crate::kv_journal::{
+  DoubleBufferedKVJournal,
+  HighWaterMarkCallback,
+  KVJournal,
+  MemMappedKVJournal,
+};
 use bd_bonjson::Value;
 use std::collections::HashMap;
 use std::path::Path;
